rag/requirements.txt
# Core dependencies
fastapi==0.112.2
uvicorn==0.30.6
requests==2.32.3
python-dotenv==1.0.1
aiofiles==24.1.0

# RAG dependencies (sentence embeddings)
sentence-transformers==3.0.1
faiss-cpu>=1.7.4

# SPARQL/RDF dependencies
rdflib==7.0.0
SPARQLWrapper==2.0.0

# LLM API dependencies (Gemini only)
google-generativeai==0.8.3

